move renameEntity, renameEntityType
